% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naturoute-package.R
\docType{package}
\name{naturoute-package}
\alias{naturoute}
\alias{naturoute-package}
\title{naturoute: Routing through natural spaces}
\description{
Routing through natural spaces.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/naturoute}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/naturoute/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
