% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{nr_prepare_data}
\alias{nr_prepare_data}
\title{Prepare data from an Open Street Map (OSM) file}
\usage{
nr_prepare_data(osm_file, osm_boundary_id, results_dir)
}
\arguments{
\item{osm_file}{A character string giving the path to an OSM file. The
function will work most efficiently if this in in \code{.pbf} format.}

\item{osm_boundary_id}{The OSM identifier of a bounding polygon for the
region of interest.}

\item{results_dir}{Directory in which processed OSM extracts are to be saved.}
}
\value{
A character vector of two elements containing full file paths to data
on the OSM network and natural spaces.
}
\description{
Prepare data from an Open Street Map (OSM) file
}
